version00/agents/requirements.txt
# Network AIOps Agent Dependencies

# LLM Provider
google-generativeai>=0.8.0

# Neo4j Driver
neo4j>=5.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: For future agents
# langgraph>=0.0.20
# langchain>=0.1.0
# langchain-neo4j>=0.0.5
